pyproject.toml
[project]
name = "hybridrag"
version = "0.1.0"
description = "LightRAG + DeepLake Hybrid RAG System"
authors = [
    {name = "HybridRAG Team"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "lightrag-hku>=0.1.0",
    "deeplake>=4.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "colorama>=0.4.6",
    "numpy>=1.24.0",
    # Additional dependencies that might be needed for the scripts
    "asyncio-throttle>=1.0.0",
    "tqdm>=4.65.0",
    # PromptChain for query interface
    "promptchain",
    "litellm>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

# No build system needed for script-only project

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.uv.sources]
promptchain = { git = "https://github.com/gyasis/PromptChain.git", branch = "main" }

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["hybridrag"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false